[UMP-v1.1] ALSA: ump: Support endpoint and function block parsing
This adds the basic support for UMP Endpoint and UMP Function Block
parsing.  The commit provides a new helper function to perform the
query of the UMP Endpoint information and builds up the UMP blocks per
each UMP Function Block information.  For the communication via a
given UMP Endpoint, it prepares a special out-of-bounce buffer and
performs the inquiry and waits for the reply with it.

Signed-off-by: Takashi Iwai <tiwai@suse.de>