ALSA: seq: Bind UMP device
Create a new ALSA sequencer client for the kernel UMP object,
snd-seq-ump-client.  It's a UMP version of snd-seq-midi driver, where
this creates a sequencer client per UMP endpoint which contains
(fixed) 16 ports.

The UMP packet group is updated upon delivery depending on the target
sequencer port (which corresponds to the actual UMP group).

Signed-off-by: Takashi Iwai <tiwai@suse.de>