ALSA: usb-audio: Get UMP EP name string from USB interface
USB descriptor may provide a nicer name for USB interface, and we may
take it as the UMP Endpoint name.

Also, fill the UMP block product_id field from the iProduct string of
the USB device descriptor, too.  It's not unique per interface, but
this is likely what we'd need in most cases.

Signed-off-by: Takashi Iwai <tiwai@suse.de>