[UMP-v1.1] ALSA: ump: Handle dynamic change of FB info
This patch extends the UMP core to handle the dynamic UMP FB info
change via UMP Stream messages.  The new UMP spec allows to update the
FB info and change its associated groups or its activeness on the
fly.  For catching it, the UMP core needs to keep watching the
incoming UMP messages, and it's done via the new function
snd_ump_receive() that is replaced with the existing
snd_rawmidi_receive() call (e.g. called from USB-audio driver).

snd_ump_receive() parses the incoming UMP packet and handles the
Stream message accordingly.  The groups and the activeness in FB info
gets updated by the message.

For enabling the incoming UMP packet parsing, the existing
OOB-handling code in the UMP core is modified (simplified) and
extended.

Signed-off-by: Takashi Iwai <tiwai@suse.de>