ALSA: ump: Redirect rawmidi substream access via own helpers
This is a code refactoring for abstracting the rawmidi access to the
UMP's own helpers.  It's a preliminary work for the later code
refactoring of the UMP layer.

Until now, we access to the rawmidi substream directly from the
driver, but after this change, the driver is supposed to access via
the newly introduced snd_ump_ops and receive/transmit via
snd_ump_receive() and snd_ump_transmit() helpers.  As of this commit,
those are merely wrappers for the rawmidi substream, and no much
function change is seen here.

Signed-off-by: Takashi Iwai <tiwai@suse.de>