ALSA: rawmidi: Skip UMP devices at SNDRV_CTL_IOCTL_RAWMIDI_NEXT_DEVICE
Applications may look for rawmidi devices with the ioctl The
SNDRV_CTL_IOCTL_RAWMIDI_NEXT_DEVICE, and returning a UMP device from
this ioctl may confuse the existing applications that support only the
legacy rawmidi.

This patch changes the code to skip the UMP devices from the lookup
for avoiding the confusion, and introduces a new ioctl to look for the
UMP devices instead.

Signed-off-by: Takashi Iwai <tiwai@suse.de>