ALSA: seq: Add port disabled flag
This extends the ALSA sequencer port capability bit to indicate the
"disabled" flag.  When this flag is set, it's essentially invisible,
and doesn't appear in the port query ioctls, while the direct access
and the connection to this port are still allowed.  The disabled state
can be flipped dynamically, so that it can be visible at any time
later.

Signed-off-by: Takashi Iwai <tiwai@suse.de>