ALSA: seq: Support client->port connection
This is an extension to ALSA sequencer infrastructure to support the
client->port connection.  Now the seq core allows to create a special
port with the capability bit SNDRV_SEQ_PORT_CAP_BROADCAST, and the
port will behave specially at reading from this port via subscription
to broadcast the events sent to any ports; that is, it'll be a sort of
catch-all port.

There can be only one broadcast port per client.

Signed-off-by: Takashi Iwai <tiwai@suse.de>